# ayunis-core-frontend/src/shared/api/openapi-schema.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/invites": {
      "post": {
        "description": "Send an invitation to a user to join an organization with a specific role",
        "operationId": "InvitesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInviteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The invite has been successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateInviteResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid invite data"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Create a new invite",
        "tags": [
          "invites"
        ]
      },
      "get": {
        "description": "Retrieve all invites for the organization with calculated status and sent date",
        "operationId": "InvitesController_getInvites",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns all invites for the organization",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InviteResponseDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get all invites for current user's organization",
        "tags": [
          "invites"
        ]
      }
    },
    "/invites/{token}": {
      "get": {
        "description": "Retrieve invite details including organization name by token",
        "operationId": "InvitesController_getInviteByToken",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "description": "Token of the invited user",
            "schema": {
              "example": "1234567890",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the invite with organization details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteDetailResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Invite not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get a single invite by token",
        "tags": [
          "invites"
        ]
      }
    },
    "/invites/accept": {
      "post": {
        "description": "Accept an invitation using the JWT token",
        "operationId": "InvitesController_acceptInvite",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInviteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The invite has been successfully accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptInviteResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid invite token or invite already accepted/expired"
          },
          "404": {
            "description": "Invite not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Accept an invite",
        "tags": [
          "invites"
        ]
      }
    },
    "/invites/{id}": {
      "delete": {
        "description": "Delete an invitation (only allowed by the user who created it)",
        "operationId": "InvitesController_deleteInvite",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The UUID of the invite to delete",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The invite has been successfully deleted"
          },
          "403": {
            "description": "Unauthorized to delete this invite"
          },
          "404": {
            "description": "Invite not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Delete an invite",
        "tags": [
          "invites"
        ]
      }
    },
    "/users": {
      "get": {
        "description": "Retrieve all users that belong to the current authenticated user's organization. Returns user information without sensitive data like password hashes.",
        "operationId": "UserController_getUsersInOrganization",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successfully retrieved users in the organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsersListResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "User not authenticated"
          },
          "500": {
            "description": "Internal server error occurred while retrieving users"
          }
        },
        "summary": "Get users in current organization",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}/role": {
      "patch": {
        "description": "Update the role of a user. You cannot update your own role.",
        "operationId": "UserController_updateUserRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID to update role for",
            "schema": {
              "format": "uuid",
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "New role information",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User role successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid role value"
          },
          "401": {
            "description": "User not authenticated or trying to update own role"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error occurred while updating user role"
          }
        },
        "summary": "Update user role",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}": {
      "delete": {
        "description": "Delete a user by their ID. Only users within the same organization can be deleted.",
        "operationId": "UserController_deleteUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID to delete",
            "schema": {
              "format": "uuid",
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User successfully deleted"
          },
          "401": {
            "description": "User not authenticated"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error occurred while deleting user"
          }
        },
        "summary": "Delete a user",
        "tags": [
          "Users"
        ]
      }
    }
  },
  "info": {
    "title": "Ayunis Core API",
    "description": "The Ayunis Core API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateInviteDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email address of the person to invite",
            "example": "user@example.com"
          },
          "role": {
            "type": "string",
            "description": "Role to assign to the invited user",
            "enum": [
              "admin",
              "user"
            ],
            "example": "user"
          }
        },
        "required": [
          "email",
          "role"
        ]
      },
      "CreateInviteResponseDto": {
        "type": "object",
        "properties": {
          "inviteToken": {
            "type": "string",
            "description": "JWT token for the invite",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "inviteToken"
        ]
      },
      "InviteResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier of the invite",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "description": "Email address of the invited user",
            "example": "user@example.com"
          },
          "role": {
            "type": "string",
            "description": "Role assigned to the invited user",
            "enum": [
              "admin",
              "user"
            ],
            "example": "user"
          },
          "status": {
            "type": "string",
            "description": "Current status of the invite",
            "enum": [
              "pending",
              "accepted",
              "expired"
            ],
            "example": "pending"
          },
          "sentDate": {
            "format": "date-time",
            "type": "string",
            "description": "Date when the invite was sent",
            "example": "2023-12-01T10:00:00Z"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string",
            "description": "Date when the invite expires",
            "example": "2023-12-08T10:00:00Z"
          },
          "acceptedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Date when the invite was accepted (if applicable)",
            "example": "2023-12-02T15:30:00Z"
          }
        },
        "required": [
          "id",
          "email",
          "role",
          "status",
          "sentDate",
          "expiresAt"
        ]
      },
      "InviteDetailResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier of the invite",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "description": "Email address of the invited user",
            "example": "user@example.com"
          },
          "role": {
            "type": "string",
            "description": "Role assigned to the invited user",
            "enum": [
              "admin",
              "user"
            ],
            "example": "user"
          },
          "status": {
            "type": "string",
            "description": "Current status of the invite",
            "enum": [
              "pending",
              "accepted",
              "expired"
            ],
            "example": "pending"
          },
          "sentDate": {
            "format": "date-time",
            "type": "string",
            "description": "Date when the invite was sent",
            "example": "2023-12-01T10:00:00Z"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string",
            "description": "Date when the invite expires",
            "example": "2023-12-08T10:00:00Z"
          },
          "acceptedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Date when the invite was accepted (if applicable)",
            "example": "2023-12-02T15:30:00Z"
          },
          "organizationName": {
            "type": "string",
            "description": "Name of the organization",
            "example": "Acme Corporation"
          }
        },
        "required": [
          "id",
          "email",
          "role",
          "status",
          "sentDate",
          "expiresAt",
          "organizationName"
        ]
      },
      "AcceptInviteDto": {
        "type": "object",
        "properties": {
          "inviteToken": {
            "type": "string",
            "description": "JWT token from the invite",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "userName": {
            "type": "string",
            "description": "Name of the user accepting the invite",
            "example": "John Doe"
          },
          "password": {
            "type": "string",
            "description": "Password of the user accepting the invite",
            "example": "password"
          },
          "passwordConfirm": {
            "type": "string",
            "description": "Confirmation of the password of the user accepting the invite",
            "example": "password"
          }
        },
        "required": [
          "inviteToken",
          "userName",
          "password",
          "passwordConfirm"
        ]
      },
      "AcceptInviteResponseDto": {
        "type": "object",
        "properties": {
          "inviteId": {
            "type": "string",
            "description": "ID of the accepted invite",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "description": "Email of the user who accepted the invite",
            "example": "user@example.com"
          },
          "orgId": {
            "type": "string",
            "description": "Organization ID the user was invited to",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "format": "uuid"
          }
        },
        "required": [
          "inviteId",
          "email",
          "orgId"
        ]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "User unique identifier",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "User name",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "description": "User email address",
            "example": "john.doe@example.com"
          },
          "role": {
            "type": "string",
            "description": "User role",
            "enum": [
              "admin",
              "user"
            ],
            "example": "user"
          },
          "orgId": {
            "type": "string",
            "description": "Organization ID the user belongs to",
            "example": "123e4567-e89b-12d3-a456-426614174001",
            "format": "uuid"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Date when the user was created",
            "example": "2024-01-15T10:30:00Z"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "role",
          "orgId",
          "createdAt"
        ]
      },
      "UsersListResponseDto": {
        "type": "object",
        "properties": {
          "users": {
            "description": "List of users in the organization",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserResponseDto"
            }
          }
        },
        "required": [
          "users"
        ]
      },
      "UpdateUserRoleDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "description": "New role for the user",
            "enum": [
              "admin",
              "user"
            ],
            "example": "admin"
          }
        },
        "required": [
          "role"
        ]
      }
    }
  }
}